Fix GeoJSONExporterTest
Turns out that the GeographySpecification params
were flipped, meaning the exporter returned an
empty set. I thought the empty set was expected by
the test but I don't think this any more.

Had to add an AttributeSpecification to get it to
run and then we got some real results out of it!

Asserting against those real results is very
unwieldy and should be mocked, as well as using a
nicer JSON library, but we'll do that in a future
commit. For now we just check against the label
of the first feature.

This should give us enough structure to make our
CSV exporter testable.