Add variable number of attributes to CSVExporter
To do this we've broken out our logic into two
fns — `lineariseGeographies` which unwinds all the
nesting into a key-value map, and then
`tabulateGeographyMap` which sets that map in a
list according to another list of attributes.

That attribute list is generated with another fn
that runs through the datasetSpecification and
pulls out all the attributes.

This might seem unnecessarily complicated — why not
just make lineariseGeographies return a list in the
same order as the attributes. Which might be true,
but I'm not totally convinced that the code should
rely on order. It seems like somewhere bugs could
easily creep in. And in any case, it gives us a
bit more fnality out of the class for little cost.