Allow specifying inputAttributes with query
So instead of having to specify the whole attribute
including the metadata, our code just takes the
attributeLabel and providerLabel and pulls the
required attribute out of the database for us.

Encountering some issues with testing code that
interacts with the database or the static utils,
which means a lot of this isn't testable. We'll
manage it by extracting that code for now, but we
should come up with a better solution long-term.