Clear TimedValues in DB between tests
Involves creating an AbstractTest with our DB
clearing code in it. This increases the test
run time a bit, sadly. We'll work on that.

The idea is to progress through the tables and
clear them between tests, fixing failures as we go,
then we can be sure our tests don't rely on anything
in the DB.