Make all *Utils sessions transient
Keeping them around adds global state which makes
a bunch of things harder. For instance, clearing
the DB between tests.

Had to amend a test to make it work. Initially
looked at fixing the fnality of `testGetUnknownGeographyType`
but it didn't seem to do anything useful so I've
just removed it. We'll see if I end up needing
to restore it...