Add `withSession(Consumer)` & test
See next commit for its utility. Also add some
tests for both versions of the fn — because we want
to use it for testing(? or do we) and that means
we should be sure it works. Lambda fns come with
their own failure mode in that the fn can sometimes
not be called, so your asserts don't run and your
tests pass — this is obviously important to guard
against.

God dang we're really going down the garden path
here.