Re-start the session-factory in tests
State being preserved between tests in these
singletons is causing issues. We need to be sure
that the sessions are all killed between tests,
because we might want to preserve them in other
normal-running situations. But not in tests.

Necessitates a change to the GeographyTests which
previously relied on a session being preserved
between tests.