Add global DB session
Having researched this a bit more, I've refined my
understanding of how long the Session obj is
supposed to live for. I think for us it should
be the duration of the execution (that is, the
parsing of the JSON specification file).

That doesn't mean we can be lazy and keep it around
for the duration of our code running, however — as
it is currently and continues to be at present.
Both because each test is a new execution and should
have a fresh session — and also because a 3rd party
may want to use our code and keeping the session
in memory forever may be unexpected to them.

I think in reality it should be explicitly started
at the beginning of a 'job', and then explicitly
closed at the end of that job. It would be nice
to enforce that via passing the session around, but
realistically we're just going to have to remember
to open and close it at the right moments.

That change comes later though.