Don't re-fetch geographyType
geographyType is already the instance that method
is trying to return (but failing to do so).

The method `getGeographyTypeByLabel` needs a look at.
It's doing something odd that isn't what the method
signature suggests. I think it relies on the fact
that the session has already been loaded with
what it needs, but I'm not 100%.

In any case, at this point in the code, in the case
where the record doesn't exist and needs to be
created — it doesn't work.