Simplify getGeographyTypeByLabel behaviour
Previously this exhibited strange behaviour with
records that were recently saved, that I've not
managed to figure out.

The amendment is due to the fact that the code is
equivalent to `.get` and this doc string from the
`.load` docs:

```
You should not use this method to determine if an instance exists (use <tt>get()</tt>
instead). Use this only to retrieve an instance that you assume exists, where non-existence
would be an actual error.
```