Clear non-default geography types
This is getting a bit messy, but I can't think of
a better way of doing it within these strictures.

Ideally we'd execute each test in a transaction
and then roll it back, but Hibernate doesn't support
nested transactions or emulation via savepoints so
unless we decide to do all our transaction handling
ourselves we're doomed to this. I guess that _is_
an alternative we could consider...

Alternatively we could create the database new each
time and re-initialise our default tables, but that's
a lot of data and it does take quite a while to insert.

We could use table templates but that would involve
giving the tests permission to create/drop databases and
that seems risky.

My initial idea was to trash the geography_object
table between tests and make all tests set up their
own geographies. But the current default dataset
(lsoas, msoas etc) is pretty wedded into those tests
and it turned into a lot of work to generalise them.
I guess that's another thing to consider, that may be
worth the trouble.