Factor Attributes handling to top layer in GeoExp
The idea here is to end up with two methods, one
of which writes with attributes and the other of
which writes with fields. To do this we remove
all notion of attribute from all fns except the
specific attribute writing one.

Next we'll use this to implement the fields one.