Add special exception for incomputable fields
Sometimes we end up not having the right data in
the database and it's annoying but it shouldn't
prevent us from proceeding. For this situation we
create a new checked exception that we throw for the
exporter to recover from & insert a null-like
value instead.

We're coming up against annoying deserialization
issues where we can't use variable initialisers
or our constructors for setup... at the moment
we just lazily set them up but we should probably
come up with a better solution.