Change ImportCacheMarkers to DatabaseJournal
The larger concept being introduced here is that
every change to the database should be accompanied
by a journal entry that identifies that change.
This way we can reason about the state of the
database and not re-perform database work if we
don't need to.

This concept can be extended to other database work
should we need to do so (though I'm not totally
convinced we should at present).