Add clearDatabaseCache property
This replaces the doImport flag. Now imports will
always be done if they're needed, but you can set
a flag to clear the whole database before the job,
so you can effectively 'refresh' or 'make clean'
instead.

We extract the code we used in the tests for this
purpose into a new utilities class.