Replace JsonPath with JsonAssert
Was struggling with a neat way of dealing with
slightly differing JSON formats in the integration
tests and decided the current approach wasn't
working but that I _really_ didn't fancy going
through with org.json.simple.

Instead, we're using JsonAssert, which is more
verbose but pays off in terms of maintainability
and precision. Note that JsonAssert has a neat
feature where it just ignores keys you don't
specify if you pass `false` as the last param, so
you don't have to repeat yourself too much when
it comes to coords, etc.