Allow only one subjectMatcher rule
Rename to matchRule for clarity and only allow one.
As far as I know, we've never actually needed or
used multiple matchers, and now without the criteria
API it's annoying to do so I removed it.

Though, perhaps more annoyingly we'll have to change
all the specfiles now... oh well. Probably should
make a habit of providing less powerful interfaces
as a rule.