Repair BasicPredefinedField unit tests
Wanted to reuse the example predefined-field in an
integration test but needed it to use the right
attributes from the actual dataset in that case.
This meant changing the unit test — but because
the TestFactory appends `_label` to all labels,
I had to change it to make it work with the
attributes...

This way is better anyway.