Add abstract importer for geotools-based importers
The best way I can think of to make this the most
testable possible. We’re never going to want to
test a live postgres server and won’t have much
fun setting up fake test postgres servers to test
against — but we can create a generic geotools
importer which we can test thoroughly and then
test the config-y bits of the child importers.

This is ugly as heck now but I’m hoping to cut it
back now the tests are passing.